[Project] Extract type definitions from 'ORKDefines.h' into the new 'ORKTypes.h' file
