+++ ResearchKit/ResearchKit.h
@@ -29,7 +29,7 @@
  */
 
 
-#import <ResearchKit/ORKDefines.h>
+#import <ResearchKit/ORKTypes.h>
 
 #import <ResearchKit/ORKTask.h>
 #import <ResearchKit/ORKOrderedTask.h>
